Moving around Old code to make way for my new restructuring
