+++ OrginalCode/README.md
@@ -1,0 +1,4 @@
+# Shear Stress Bioreactor - Src
+This directory contains the source code of the program. Code for the hardware controls is in `controls/` and code for the interface is in `interface/`
+
+The program entry point, `main.cpp` is located in this directory.
