+++ OrginalCode/controls/utils.cpp
