+++ OrginalCode/interface/README.md
@@ -1,0 +1,5 @@
+# Shear Stress Bioreactor - Interface
+This folder contains the software which provides the user interface of the project.
+
+# On building
+Running the [Lua](https://lua.org) script `buildHtml.lua` in the root of the repository build the html.hpp file, and also prints to the console the required `server->on` definitions, which can be copy-pasted into `gui.cpp`
